% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_title.R
\name{check_title_length}
\alias{check_title_length}
\title{Check title suitability}
\usage{
check_title_length(title)
}
\arguments{
\item{title}{The article title: a short string}
}
\value{
An output describing the suitability of the title for research discovery based on its length and the number of non-stop words
}
\description{
Check given title for an article to assess how discoverable it is based on its length and proportion of words that are non-stop words
}
\examples{
title <- "A methodology for systematic mapping in environmental sciences"
check <- check_title_length(title)
check;
}
