% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litsearchr.functions.R
\name{fakerake}
\alias{fakerake}
\title{Functions from litsearchr (not yet on CRAN)
Quick keyword extraction}
\usage{
fakerake(text, stopwords, min_n = 2, max_n = 5)
}
\arguments{
\item{text}{A string object to extract terms from}

\item{stopwords}{A character vector of stop words to remove}

\item{min_n}{Numeric: the minimum length ngram to consider}

\item{max_n}{Numeric: the maximum length ngram to consider}
}
\value{
A character vector of potential keywords
}
\description{
Extracts potential keywords from text separated by stop words
}
