% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_keywords.R
\name{format_keywords}
\alias{format_keywords}
\title{Format input keywords}
\usage{
format_keywords(keywords, sep = ";")
}
\arguments{
\item{keywords}{The article keywords: a vector of strings}

\item{sep}{Character that separates keywords in a single string}
}
\value{
A vector of lowercase keywords
}
\description{
Convert string of keywords with separator into a vector
}
\examples{
keywords <- c("Systematic mapping; 
  Evidence-based environmental management; 
  Systematic evidence synthesis; 
  Evidence review; 
  Knowledge gaps; 
  Knowledge clusters")
newkeywords <- format_keywords(keywords, sep = ";")
newkeywords;
}
