% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litsearchr.functions.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Retrieve stop words for a given language}
\usage{
get_stopwords(language = "English")
}
\arguments{
\item{language}{A character vector containing the name of the language for which to retrieve stop words. Defaults to "English"}
}
\value{
Returns a character vector of stop words.
}
\description{
This function retrieves stop words to use for a specified language.
}
\examples{
get_stopwords("English")
}
